% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_test.R
\name{contingency_test}
\alias{contingency_test}
\title{Test Independence of Factors in a 2x2 Contingency Table}
\usage{
contingency_test(tab)
}
\arguments{
\item{tab}{An integer matrix with two columns and two rows, where the rows
represent a factor with two possible values and the columns represent
another factor with two possible values, and each cell gives the number
of observations that exhibit the factor values represented by that
column and that row}
}
\value{
An object of class "ContingencyTableTests", consisting of a
dataframe with three observations of two variables, "Quantity" and
"Value". The "Quantity" variable names the quantity returned, and its
observations will be "Fisher's Exact Text p-value", "G-test p-value",
and "Bayes Factor". The "Value" variable gives the p-values and Bayes
factor value for those tests.\preformatted{The result object will also have a "contingency_table" attribute whose
value is equal to the "tab" parameter.
}
}
\description{
There are a number of statistical techniques that may be used to test
whether two factors in a contingency table are independent or related.
The G-test is a frequentist test of the hypothesis that the factors are
related, appropriate when you have at least 1,000 observations; with a
small number of observations, Fisher's exact test is more accurate
(McDonald 2014, 68-85). For these test, you choose a significance level
before running the test, commonly 0.05. The test will return a p-value,
and if that value is below the chosen significance level, then you can
reject the null hypothesis that the factors are independent of each other.
}
\details{
We can use Bayesian analysis and generate a Bayes Factor comparing the
hypotheses that the factors are related and that they're independent using
the procedure in Gunel and Dickey (1974). This Bayes Factor will tell you
how much more likely it is to have observed your data under the hypothesis
that the factors are related than under the hypothesis that they are
independent. Although there are not strict standards regarding the exact
treatment of specific Bayes factor values, generally a Bayes factor of 10
or greater is considered strong evidence of a hypothesis
(Lee and Wagenmakers 2014, 105).
}
\references{
Gunel, Erdogan and James Dickey. 1974. "Bayes Factors for
Independence in Contingency Tables". Biometrika 61(3): 545-557.

Lee, Michael D. and Eric-Jan Wagenmakers. 2014. Bayesian Cognitive Modeling:
A Practical Course. Cambridge University Press.

McDonald, J.H. 2014. Handbook of Biological Statistics (3rd ed.).
Sparky House Publishing, Baltimore, Maryland.
}
\seealso{
print.ContingencyTableTests
}
