% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_test.R
\name{print.ContingencyTableTests}
\alias{print.ContingencyTableTests}
\title{Print the results of contingency_test()}
\usage{
\method{print}{ContingencyTableTests}(
  x,
  ...,
  simplify = TRUE,
  stars = TRUE,
  show_table = TRUE,
  show_BF_guide = FALSE
)
}
\arguments{
\item{x}{A \code{ContingencyTableTests} object
(see \code{\link{contingency_test}})}

\item{...}{Arguments passed to or from other methods (currently unused)}

\item{simplify}{Should values under 0.001 be shown as < 0.001 rather than
its exact value, and should values over 100 be shown as > 100?
Default is \code{TRUE}.}

\item{stars}{Should stars be appended to p-values that would be treated as
significant at various levels, and apostrophes appended to Bayes Factors
that would be treated as at least strong evidence in favor of the
hypothesis that the factors are related? The default is \code{TRUE}.
(If \code{TRUE}, a legend as to what significance or evidence strength
the glyphs signify will also be displayed).}

\item{show_table}{Should the contingency table be displayed in addition to
the results of the statistical tests? Default is \code{TRUE}.}

\item{show_BF_guide}{Should a guide to the strength of evidence provided by
different Bayes Factor values be displayed? Default is \code{FALSE}.}
}
\value{
Returns x invisibly
}
\description{
Print the results of contingency_test()
}
